Refactor the GitHub git DSL to the git DSL
