+++ source/runner/runners/_tests/vm2.test.ts
@@ -48,11 +48,8 @@ runners.forEach(run => {
       const platform = new FakePlatform()
       exec = new Executor(new FakeCI({}), platform, run.fn, config)
 
-      // platform.getPlatformGitRepresentation = async () =
-      // platform.getPlatformDSLRepresentation = jest.fn()
-
       const dsl = await jsonDSLGenerator(platform)
-      dsl.github = { pr: {} } as any
+      dsl.github = { pr: { number: 1, base: { sha: "321" }, head: { sha: "123", repo: { full_name: "123" } } } } as any
       const realDSL = await jsonToDSL(dsl)
       return contextForDanger(realDSL)
     }
